--- core-ui/src/components/Predefined/Create/ClusterAddonsConfigurations.create.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { AddonsConfigurations } from './AddonsConfigurations/AddonsConfigurations.js';
-
-export const ClusterAddonsConfigurationsCreate = ({
-  formElementRef,
-  onChange,
-  resourceType,
-  resourceUrl,
-  namespace,
-  refetchList,
-}) => (
-  <AddonsConfigurations
-    formElementRef={formElementRef}
-    onChange={onChange}
-    resourceType={resourceType}
-    resourceUrl={resourceUrl}
-    namespace={namespace}
-    refetchList={refetchList}
-  />
-);
