--- core-ui/src/components/Predefined/Details/RoleBindings/RoleBinding.details.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import { RoleBindings } from './RoleBindings.js';
-
-export const RoleBindingsDetails = DefaultRenderer => ({ ...otherParams }) => (
-  <DefaultRenderer {...otherParams} customComponents={[RoleBindings]} />
-);
