--- core-ui/src/components/Predefined/Details/Roles/ClusterRole.details.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Rules } from './Rules.js';
-
-export const ClusterRolesDetails = DefaultRenderer => ({ ...otherParams }) => {
-  return <DefaultRenderer {...otherParams} customComponents={[Rules]} />;
-};
