+++ core-ui/src/components/Predefined/Details/index.js
@@ -6,9 +6,7 @@ export * from './Function.details';
 export * from './Secret.details';
 export * from './Service.details.js';
 export * from './Application/Application.details.js';
-export * from './Roles/ClusterRole.details.js';
-export * from './Roles/Role.details.js';
-export * from './RoleBindings/ClusterRoleBinding.details.js';
-export * from './RoleBindings/RoleBinding.details.js';
+export * from './Roles/Roles.details.js';
+export * from './RoleBindings/RoleBindings.details.js';
 export * from './OAuth2Clients/OAuth2Clients.details';
 export * from './ConfigMap.details.js';
